requirements.txt
autopep8==2.0.2
glfw==2.5.6
install==1.3.5
numpy==1.24.2
Pillow==9.4.0
pycodestyle==2.10.0
PyOpenGL @ file:///C:/Users/logre/OneDrive/OpenGL/py/WorkPlace/PyOpenGL-3.1.6-cp310-cp310-win_amd64.whl
PyOpenGL-accelerate @ file:///C:/Users/logre/OneDrive/OpenGL/py/WorkPlace/PyOpenGL_accelerate-3.1.6-cp310-cp310-win_amd64.whl
tomli==2.0.1
